+++ SA5AIUC10/Src/wdt.h
@@ -11,5 +11,6 @@
 extern void wdt_init(void);
 void wdt_isr(void);
 void soc_wdt_reset_process(void);
+void wdt_reset(void);
 
 #endif /* WDT_H_ */